api/serializers.py
from rest_framework import serializers
from terremotito.models import Empresa, Intensidad, Sensor, Usuario, Zona

class EmpresaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Empresa
        fields = '__all__'

class IntensidadSerializer(serializers.ModelSerializer):
    class Meta:
        model = Intensidad
        fields = '__all__'

class SensorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Sensor
        fields = '__all__'

class UsuarioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Usuario
        fields = '__all__'

class ZonaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Zona
        fields = '__all__'

